pyproject.toml
[project]
name = "dump1090-redis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py1090>=1.1",
    "redis>=6.2.0",
    "paho-mqtt",
]
